fix label and local descriptions
