+++ frontend/src/routes/Applications/CreateApplication/Subscription/controlData/utils.js
@@ -375,83 +375,6 @@ export const getGitBranches = async (groupControlData, setLoadingState) => {
     retrieveGitDetails(null, groupControlData, setLoadingState)
 }
 
-export const setAvailableRules = (control, result) => {
-    const { loading } = result
-    const { data, i18n } = result
-    const placementRules = data
-    control.isLoading = false
-    const error = placementRules ? null : result.error
-    if (!control.available) {
-        control.available = []
-        control.availableData = []
-    }
-    if (error || placementRules) {
-        if (error) {
-            control.isFailed = true
-            control.isLoaded = true
-        } else if (placementRules) {
-            control.isLoaded = true
-
-            control.availableInfo = {}
-            const keyFn = (rule) => {
-                const ruleName = _.get(rule, 'metadata.name', '')
-                const clusterSelector = _.get(rule, 'spec.clusterSelector')
-                const clusterConditions = _.get(rule, 'spec.clusterConditions')
-                let selector = i18n('creation.app.local.clusters.only', [ruleName])
-                if (clusterSelector?.matchExpressions?.length > 0) {
-                    if (clusterSelector.matchExpressions[0]?.key !== 'local-cluster') {
-                        const getLabels = () => {
-                            return clusterSelector.matchExpressions
-                                .map(({ key, values }) => {
-                                    return `${key}=${values.join(', ')}`
-                                })
-                                .join('; ')
-                        }
-                        selector = i18n('creation.app.clusters.matching', [ruleName, getLabels()])
-                    }
-                } else if (clusterConditions && clusterConditions[0]?.type === 'ManagedClusterConditionAvailable') {
-                    selector = i18n('creation.app.clusters.all.online', [ruleName])
-                } else if (clusterSelector.matchLabels) {
-                    if (!clusterSelector.matchLabels['local-cluster']) {
-                        selector = i18n('creation.app.clusters.matching', [ruleName, clusterSelector.matchLabels[0]])
-                    }
-                }
-                control.availableInfo[ruleName] = selector
-                return ruleName
-            }
-            control.availableData = _.keyBy(placementRules, keyFn)
-            control.available = _.map(Object.values(control.availableData), keyFn).sort()
-            control.info = ''
-            const { groupControlData } = control
-
-            // if no existing placement rules...
-            const existingRuleControl = getControlByID(groupControlData, existingRuleCheckbox)
-            existingRuleControl.disabled = control.available.length === 0
-            if (control.available.length === 0) {
-                control.placeholder = i18n('creation.app.select.no.existing.placement.rule')
-                const clusterSelectorControl = getControlByID(groupControlData, clusterSelectorCheckbox)
-                clusterSelectorControl.onSelect()
-            } else {
-                control.placeholder = i18n('creation.app.settings.existingRule')
-                existingRuleControl.onSelect()
-            }
-
-            //remove default placement rule name if this is not on the list of available placements
-            //in that case the name was set by the reverse function on control initialization
-            if (control.active) {
-                if (!control.available.includes(control.active)) {
-                    control.active = null
-                } else {
-                    control.info = control.availableInfo[control.active]
-                }
-            }
-        }
-    } else {
-        control.isLoading = loading
-    }
-    return control
-}
-
 export const setAvailableNSSpecs = (control, result) => {
     const { loading } = result
     const { data } = result
